Move some functions into shared utilities
